requirements.txt
streamlit
mysql-connector-python
pydantic
pandas